Extrack error handling to CopybookResolver
